files/log-generator/requirements.txt
pyyaml>=6.0.1
faker>=22.6.0
python-dateutil>=2.8.2
